12_Certification_Challenge/ARIA/pyproject.toml
[project]
name = "aria"
version = "0.1.0"
description = "ARIA - Agentic RAG system with conversational interface for water sector regulation documents"
requires-python = ">=3.10,<3.14"
dependencies = [
    # LLM & Orchestration
    "langchain>=0.3",
    "langchain-ollama>=0.3",
    "langchain-community>=0.3",
    "langgraph>=1.0",

    # Vector Store
    "qdrant-client>=1.14",
    "langchain-qdrant>=0.2",

    # Document Processing
    "docling>=2.70",

    # Embeddings (via Ollama - no extra package needed)

    # Web Search
    "duckduckgo-search>=7.0",

    # UI
    "chainlit>=2.5",

    # Utils
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.setuptools.packages.find]
include = ["aria*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
